backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["SamerIssam <samer.salem@kpmg.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
mangum = "^0.17.0"
pytest = "^7.4.4"
httpx = "^0.26.0"
pytest-cov = "^4.1.0"
python-multipart = "^0.0.6"
boto3 = "^1.34.34"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
addopts = "-v -rsxX"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
